+++ src/main/java/com/ssafy/bango/domain/member/service/MemberService.java
@@ -1,8 +1,8 @@
 package com.ssafy.bango.domain.member.service;
 
-import com.ssafy.bango.domain.member.controller.dto.request.GetAccessTokenRequest;
-import com.ssafy.bango.domain.member.controller.dto.request.SignUpRequest;
-import com.ssafy.bango.domain.member.controller.dto.response.TokenResponse;
+import com.ssafy.bango.domain.member.dto.request.GetAccessTokenRequest;
+import com.ssafy.bango.domain.member.dto.request.SignUpRequest;
+import com.ssafy.bango.domain.member.dto.response.TokenResponse;
 import com.ssafy.bango.domain.member.dao.MemberDAO;
 import com.ssafy.bango.domain.member.dto.Member;
 import com.ssafy.bango.domain.member.dto.SocialPlatform;
